refactor(metrics): separate abstract class and it's implementation for metrics service
