+++ src/monitoring/domain/metrics.service.ts
@@ -1,0 +1,4 @@
+export abstract class MetricsService {
+  abstract incWeatherCacheHit(): void;
+  abstract incWeatherCacheMiss(): void;
+}
