+++ src/monitoring/infrastructure/prom-metrics.service.ts
@@ -1,9 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import { Counter } from 'prom-client';
 import { InjectMetric } from '@willsoto/nestjs-prometheus';
+import { MetricsService } from '../domain/metrics.service';
 
 @Injectable()
-export class MetricsService {
+export class PromMetricsService implements MetricsService {
   constructor(
     @InjectMetric('weather_cache_hit')
     private readonly weatherCacheHit: Counter,
